Add jcr repository support in storing and fetching the product images.

+++ src/main/java/com/synectiks/demo/site/controllers/AdminProductController.java
@@ -13,6 +13,7 @@ import javax.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
@@ -23,11 +24,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.synectiks.commons.entities.demo.Product;
 import com.synectiks.commons.utils.IUtils;
+import com.synectiks.demo.site.dto.CustomerDTO;
 import com.synectiks.demo.site.dto.ProductDTO;
 import com.synectiks.demo.site.repositories.ProductRepository;
 import com.synectiks.demo.site.utils.IDemoUtils;
@@ -58,10 +61,18 @@ public class AdminProductController {
 
 	private Path path;
 	@Autowired
+	private Environment env;
+	@Autowired
+	private RestTemplate rest;
+	@Autowired
 	private ProductRepository prodRepo;
 
 	@RequestMapping("/inventory/add")
-	public ModelAndView addProduct(ModelMap model) {
+	public ModelAndView addProduct(HttpServletRequest request, ModelMap model) {
+		CustomerDTO customer = IDemoUtils.validateUser(request);
+		if (IUtils.isNull(customer)) {
+			return new ModelAndView("login");
+		}
 		ProductDTO product = new ProductDTO();
 		product.setCondition("New");
 		model.addAttribute("categoryList", categories);
@@ -73,6 +84,10 @@ public class AdminProductController {
 	@RequestMapping(value = "/inventory/add", method = RequestMethod.POST)
 	public String addProductPost(@Valid @ModelAttribute("productDTO") ProductDTO productDTO,
 			BindingResult result, HttpServletRequest request, ModelMap model) {
+		CustomerDTO customer = IDemoUtils.validateUser(request);
+		if (IUtils.isNull(customer)) {
+			return "redirect:/login";
+		}
 		logger.info("Obj: " + productDTO + "\n" + model);
 		model.addAttribute("categoryList", categories);
 		if (result.hasErrors()) {
@@ -80,11 +95,14 @@ public class AdminProductController {
 		}
 		if (!IUtils.isNull(productDTO)) {
 			if (IUtils.isNullOrEmpty(productDTO.getName()))
-				result.addError(new FieldError("productDTO", "name", "Product name is null."));
+				result.addError(new FieldError("productDTO", "name",
+						"Product name is null."));
 			if (IUtils.isNullOrEmpty(productDTO.getManufacturer()))
-				result.addError(new FieldError("productDTO", "manufacturer", "Product manufacturer is null."));
+				result.addError(new FieldError("productDTO", "manufacturer",
+						"Product manufacturer is null."));
 			if (productDTO.getPrice() <= 0)
-				result.addError(new FieldError("productDTO", "price", "Product price must be > 0"));
+				result.addError(new FieldError("productDTO", "price",
+						"Product price must be > 0"));
 
 			if (result.hasErrors()) {
 				return "redirect:/admin/inventory/add";
@@ -102,6 +120,11 @@ public class AdminProductController {
 		if (image != null && !image.isEmpty()) {
 			try {
 				image.transferTo(new File(path.toString()));
+				// Save image into jcr repository
+				String nodePath = String.format(
+						IDemoUtils.JCR_IMAGE_PATH, productDTO.getCategory());
+				IDemoUtils.createFileNode(IDemoUtils.getApiUrl(
+						env, IDemoUtils.JCR_CREATE_NODE), rest, path, nodePath);
 			} catch (Exception e) {
 				e.printStackTrace();
 				throw new RuntimeException("The product image could not be saved.", e);
@@ -111,8 +134,14 @@ public class AdminProductController {
 	}
 
 	@RequestMapping("/inventory/edit/{id}")
-	public String editProduct(@PathVariable("id") String id, Model model) {
-		ProductDTO productDTO = IDemoUtils.wrapInDTO(prodRepo.findById(id), ProductDTO.class);
+	public String editProduct(@PathVariable("id") String id,
+			HttpServletRequest request, Model model) {
+		CustomerDTO customer = IDemoUtils.validateUser(request);
+		if (IUtils.isNull(customer)) {
+			return "redirect:/login";
+		}
+		ProductDTO productDTO = IDemoUtils.wrapInDTO(
+				prodRepo.findById(id), ProductDTO.class);
 		model.addAttribute("categoryList", categories);
 		model.addAttribute("product", productDTO);
 
@@ -123,6 +152,10 @@ public class AdminProductController {
 	public String editProductPost(@ModelAttribute("productDTO") ProductDTO productDTO,
 			BindingResult result, Model model, HttpServletRequest request)
 			throws RuntimeException {
+		CustomerDTO customer = IDemoUtils.validateUser(request);
+		if (IUtils.isNull(customer)) {
+			return "redirect:/login";
+		}
 		model.addAttribute("categoryList", categories);
 		if (result.hasErrors()) {
 			return "editProduct";
@@ -135,6 +168,11 @@ public class AdminProductController {
 		if (image != null && !image.isEmpty()) {
 			try {
 				image.transferTo(new File(path.toString()));
+				// Save image into jcr repository
+				String nodePath = String.format(
+						IDemoUtils.JCR_IMAGE_PATH, productDTO.getCategory());
+				IDemoUtils.createFileNode(IDemoUtils.getApiUrl(
+						env, IDemoUtils.JCR_CREATE_NODE), rest, path, nodePath);
 			} catch (Exception e) {
 				throw new RuntimeException("The product image could not be saved." + e);
 			}
@@ -147,12 +185,22 @@ public class AdminProductController {
 	@RequestMapping("/inventory/remove/{id}")
 	public String deleteProduct(@PathVariable("id") String id, Model model,
 			HttpServletRequest request) {
+		CustomerDTO customer = IDemoUtils.validateUser(request);
+		if (IUtils.isNull(customer)) {
+			return "redirect:/login";
+		}
 		String rootDirectory = request.getSession().getServletContext().getRealPath("/");
 		path = Paths.get(rootDirectory + "/resources/product_images/"
 				+ id + ".png");
 		if (Files.exists(path)) {
 			try {
 				Files.delete(path);
+				Product prod = prodRepo.findById(id);
+				// Save image into jcr repository
+				String nodePath = String.format(
+						IDemoUtils.JCR_IMAGE_PATH, prod.getCategory());
+				IDemoUtils.removeFileNode(IDemoUtils.getApiUrl(
+						env, IDemoUtils.JCR_REMOVE_NODE), rest, path, nodePath);
 			} catch (Exception e) {
 				logger.error(e.getMessage(), e);
 			}
